Added support for sizing and color today view tasks. Also added support for deleting today view tasks
Today View tasks will not appear as the color of its parent project

Today View tasks will be different sizes based on the estimated time.

If a user deletes a task, it will also be deleted from their today view.

If a user drags a task out of the today view, it will be deleted from the today view